backend/requirements.txt
backend_utils==0.2.0
Flask==3.1.0
# Pillow==9.0.1
gunicorn==23.0.0
Pillow==11.0.0
torch==2.4.0
transformers==4.44.0
